Add current prices to market item details
