+++ backend/src/main/kotlin/net/jonasmf/auctionengine/repository/rds/AuctionMarketItemDetailRepository.kt
@@ -1,98 +1,9 @@
 package net.jonasmf.auctionengine.repository.rds
 
 import org.springframework.jdbc.core.JdbcTemplate
-import org.springframework.jdbc.core.RowMapper
 import org.springframework.stereotype.Repository
 import java.sql.Date
-import java.sql.ResultSet
 import java.time.LocalDate
-import java.time.OffsetDateTime
-import java.time.ZoneOffset
-
-data class AuctionMarketItemDetailDailyRow(
-    val statDate: LocalDate,
-    val pointTimestamp: OffsetDateTime,
-    val minPrice: Long?,
-    val avgPrice: Double?,
-    val p25Price: Long?,
-    val p75Price: Long?,
-    val maxPrice: Long?,
-    val minQuantity: Long?,
-    val avgQuantity: Double?,
-    val maxQuantity: Long?,
-)
-
-data class AuctionMarketItemDetailHourlyRow(
-    val timestamp: OffsetDateTime,
-    val hourOfDay: Int,
-    val minPrice: Long?,
-    val avgPrice: Double?,
-    val p25Price: Long?,
-    val p75Price: Long?,
-    val maxPrice: Long?,
-    val totalQuantity: Long?,
-)
-
-data class AuctionMarketItemDetailPieRow(
-    val hourOfDay: Int,
-    val quantity: Long?,
-    val fraction: Double,
-)
-
-data class AuctionMarketItemHeaderRow(
-    val itemId: Int,
-    val itemName: String,
-    val itemMediaUrl: String?,
-    val qualityId: Int?,
-    val qualityType: String?,
-    val qualityName: String?,
-    val itemClassId: Int?,
-    val itemClassName: String?,
-    val itemSubclassId: Int?,
-    val itemSubclassName: String?,
-    val recipeId: Int?,
-    val recipeName: String?,
-    val recipeMediaUrl: String?,
-)
-
-data class AuctionMarketItemCraftingRow(
-    val recipeId: Int,
-    val recipeName: String,
-    val recipeMediaUrl: String?,
-    val craftedQuantity: Int,
-    val reagentCost: Long?,
-    val reagentsFullyPriced: Boolean,
-    val outputUnitPrice: Long?,
-    val profit: Long?,
-    val roiPercent: Double?,
-)
-
-data class AuctionMarketItemCraftingReagentRow(
-    val recipeId: Int,
-    val itemId: Int,
-    val name: String,
-    val mediaUrl: String?,
-    val quantity: Int,
-    val unitPrice: Long?,
-    val lineTotal: Long?,
-)
-
-data class AuctionMarketItemCraftingAnalyticsDailyRow(
-    val statDate: LocalDate,
-    val profit: Long?,
-    val roiPercent: Double?,
-    val reagentCost: Long?,
-    val outputUnitPrice: Long?,
-)
-
-data class AuctionMarketItemCraftingHeatmapRow(
-    val dayOfWeek: Int,
-    val hourOfDay: Int,
-    val profit: Double?,
-    val outputUnitPrice: Double?,
-    val roiPercent: Double?,
-    val sampleCount: Int,
-)
 
 @Repository
 class AuctionMarketItemDetailRepository(
@@ -122,7 +33,7 @@ class AuctionMarketItemDetailRepository(
             WHERE d.item_id = ?
             LIMIT 1
             """.trimIndent(),
-            headerRowMapper,
+            AuctionMarketItemDetailRowMappers.headerRowMapper,
             itemId,
         ).firstOrNull()
 
@@ -172,14 +83,29 @@ class AuctionMarketItemDetailRepository(
                 )
             }
         return jdbcTemplate
-            .query(sql, snapshotRowMapper, *params)
+            .query(sql, AuctionMarketItemDetailRowMappers.snapshotRowMapper, *params)
             .firstOrNull() ?: (null to null)
     }
 
-    private val snapshotRowMapper =
-        RowMapper { rs: ResultSet, _: Int ->
-            rs.getNullableLong("price") to rs.getNullableLong("qty")
-        }
+    fun loadCurrentListings(
+        connectedRealmId: Int,
+        itemId: Int,
+        variant: Boolean,
+        bonusKey: String,
+        modifierKey: String,
+        petSpeciesId: Int,
+    ): List<AuctionMarketItemCurrentListingRow> {
+        val (sql, params) =
+            buildCurrentListingsSqlAndArgs(
+                connectedRealmId,
+                itemId,
+                variant,
+                bonusKey,
+                modifierKey,
+                petSpeciesId,
+            )
+        return jdbcTemplate.query(sql, AuctionMarketItemDetailRowMappers.currentListingRowMapper, *params)
+    }
 
     fun loadDailySeries(
         connectedRealmId: Int,
@@ -202,7 +128,7 @@ class AuctionMarketItemDetailRepository(
                 modifierKey,
                 petSpeciesId,
             )
-        return jdbcTemplate.query(sql, dailyRowMapper, *params)
+        return jdbcTemplate.query(sql, AuctionMarketItemDetailRowMappers.dailyRowMapper, *params)
     }
 
     fun loadHourlySeries(
@@ -226,7 +152,7 @@ class AuctionMarketItemDetailRepository(
                 modifierKey,
                 petSpeciesId,
             )
-        return jdbcTemplate.query(sql, hourlyRowMapper, *params)
+        return jdbcTemplate.query(sql, AuctionMarketItemDetailRowMappers.hourlyRowMapper, *params)
     }
 
     fun loadQuantityPie(
@@ -248,7 +174,7 @@ class AuctionMarketItemDetailRepository(
                 modifierKey,
                 petSpeciesId,
             )
-        return jdbcTemplate.query(sql, pieRowMapper, *params)
+        return jdbcTemplate.query(sql, AuctionMarketItemDetailRowMappers.pieRowMapper, *params)
     }
 
     fun loadCraftings(
@@ -282,7 +208,7 @@ class AuctionMarketItemDetailRepository(
                 preferredRecipeId,
                 localeColumnSuffix,
             )
-        return jdbcTemplate.query(sql, craftingRowMapper, *params)
+        return jdbcTemplate.query(sql, AuctionMarketItemDetailRowMappers.craftingRowMapper, *params)
     }
 
     fun loadCraftingReagents(
@@ -365,111 +291,25 @@ class AuctionMarketItemDetailRepository(
                 *recipeIds.toTypedArray(),
                 *recipeIds.toTypedArray(),
             )
-        return jdbcTemplate.query(sql, craftingReagentRowMapper, *params)
+        return jdbcTemplate.query(sql, AuctionMarketItemDetailRowMappers.craftingReagentRowMapper, *params)
     }
 
-    private val headerRowMapper =
-        RowMapper { rs: ResultSet, _: Int ->
-            AuctionMarketItemHeaderRow(
-                itemId = rs.getInt("item_id"),
-                itemName = rs.getString("item_name"),
-                itemMediaUrl = rs.getString("item_media_url"),
-                qualityId = rs.getNullableInt("quality_id"),
-                qualityType = rs.getString("quality_type"),
-                qualityName = rs.getString("quality_name"),
-                itemClassId = rs.getNullableInt("item_class_id"),
-                itemClassName = rs.getString("item_class_name"),
-                itemSubclassId = rs.getNullableInt("item_subclass_id"),
-                itemSubclassName = rs.getString("item_subclass_name"),
-                recipeId = rs.getNullableInt("recipe_id"),
-                recipeName = rs.getString("recipe_name"),
-                recipeMediaUrl = rs.getString("recipe_media_url"),
-            )
-        }
-
-    private val dailyRowMapper =
-        RowMapper { rs: ResultSet, _: Int ->
-            val d = rs.getObject("stat_date", LocalDate::class.java)
-            AuctionMarketItemDetailDailyRow(
-                statDate = d,
-                pointTimestamp = d.atStartOfDay(ZoneOffset.UTC).toOffsetDateTime(),
-                minPrice = rs.getNullableLong("min_price"),
-                avgPrice = rs.getNullableDouble("avg_price"),
-                p25Price = rs.getNullableLong("p25_price"),
-                p75Price = rs.getNullableLong("p75_price"),
-                maxPrice = rs.getNullableLong("max_price"),
-                minQuantity = rs.getNullableLong("min_quantity"),
-                avgQuantity = rs.getNullableDouble("avg_quantity"),
-                maxQuantity = rs.getNullableLong("max_quantity"),
-            )
-        }
-
-    private val hourlyRowMapper =
-        RowMapper { rs: ResultSet, _: Int ->
-            val ts = rs.getTimestamp("timestamp").toInstant().atOffset(ZoneOffset.UTC)
-            AuctionMarketItemDetailHourlyRow(
-                timestamp = ts,
-                hourOfDay = rs.getInt("hour_of_day"),
-                minPrice = rs.getNullableLong("min_price"),
-                avgPrice = rs.getNullableDouble("avg_price"),
-                p25Price = rs.getNullableLong("p25_price"),
-                p75Price = rs.getNullableLong("p75_price"),
-                maxPrice = rs.getNullableLong("max_price"),
-                totalQuantity = rs.getNullableLong("total_quantity"),
-            )
-        }
-
-    private val pieRowMapper =
-        RowMapper { rs: ResultSet, _: Int ->
-            AuctionMarketItemDetailPieRow(
-                hourOfDay = rs.getInt("hour_of_day"),
-                quantity = rs.getNullableLong("quantity"),
-                fraction = rs.getDouble("fraction"),
-            )
-        }
-
-    private val craftingRowMapper =
-        RowMapper { rs: ResultSet, _: Int ->
-            AuctionMarketItemCraftingRow(
-                recipeId = rs.getInt("recipe_id"),
-                recipeName = rs.getString("recipe_name") ?: "",
-                recipeMediaUrl = rs.getString("recipe_media_url"),
-                craftedQuantity = rs.getInt("crafted_quantity"),
-                reagentCost = rs.getNullableLong("reagent_cost"),
-                reagentsFullyPriced = rs.getBoolean("reagents_fully_priced"),
-                outputUnitPrice = rs.getNullableLong("output_unit_price"),
-                profit = rs.getNullableLong("profit"),
-                roiPercent = rs.getNullableDouble("roi_percent"),
-            )
-        }
-
-    private val craftingReagentRowMapper =
-        RowMapper { rs: ResultSet, _: Int ->
-            AuctionMarketItemCraftingReagentRow(
-                recipeId = rs.getInt("recipe_id"),
-                itemId = rs.getInt("item_id"),
-                name = rs.getString("name") ?: "",
-                mediaUrl = rs.getString("media_url"),
-                quantity = rs.getInt("quantity"),
-                unitPrice = rs.getNullableLong("unit_price"),
-                lineTotal = rs.getNullableLong("line_total"),
-            )
-        }
-
-    private fun ResultSet.getNullableInt(column: String): Int? {
-        val v = getInt(column)
-        return if (wasNull()) null else v
-    }
-
-    private fun ResultSet.getNullableLong(column: String): Long? {
-        val v = getLong(column)
-        return if (wasNull()) null else v
-    }
-
-    private fun ResultSet.getNullableDouble(column: String): Double? {
-        val v = getDouble(column)
-        return if (wasNull()) null else v
-    }
+    internal fun buildCurrentListingsSqlAndArgs(
+        connectedRealmId: Int,
+        itemId: Int,
+        variant: Boolean,
+        bonusKey: String,
+        modifierKey: String,
+        petSpeciesId: Int,
+    ): Pair<String, Array<Any?>> =
+        AuctionMarketItemDetailSql.buildCurrentListingsSqlAndArgs(
+            connectedRealmId,
+            itemId,
+            variant,
+            bonusKey,
+            modifierKey,
+            petSpeciesId,
+        )
 
     internal fun buildDailySqlAndArgs(
         connectedRealmId: Int,
@@ -480,143 +320,17 @@ class AuctionMarketItemDetailRepository(
         bonusKey: String,
         modifierKey: String,
         petSpeciesId: Int,
-    ): Pair<String, Array<Any?>> {
-        val sql =
-            if (variant) {
-                """
-                WITH RECURSIVE date_spine AS (
-                    SELECT ? AS stat_date
-                    UNION ALL
-                    SELECT DATE_ADD(stat_date, INTERVAL 1 DAY)
-                    FROM date_spine
-                    WHERE stat_date < ?
-                ),
-                priced AS (
-                    SELECT
-                        v.date,
-                        v.price,
-                        v.quantity,
-                        PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY v.price) OVER (
-                            PARTITION BY v.connected_realm_id, v.item_id, v.bonus_key, v.modifier_key, v.pet_species_id, v.date
-                        ) AS p25_price,
-                        PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY v.price) OVER (
-                            PARTITION BY v.connected_realm_id, v.item_id, v.bonus_key, v.modifier_key, v.pet_species_id, v.date
-                        ) AS p75_price
-                    FROM v_auction_house_prices v
-                    WHERE v.connected_realm_id = ?
-                      AND v.item_id = ?
-                      AND v.date BETWEEN ? AND ?
-                      AND v.bonus_key <=> ?
-                      AND v.modifier_key <=> ?
-                      AND v.pet_species_id = ?
-                ),
-                daily_agg AS (
-                    SELECT
-                        date AS stat_date,
-                        MIN(price) AS min_price,
-                        AVG(price) AS avg_price,
-                        MIN(p25_price) AS p25_price,
-                        MIN(p75_price) AS p75_price,
-                        MAX(price) AS max_price,
-                        MIN(quantity) AS min_quantity,
-                        AVG(quantity) AS avg_quantity,
-                        MAX(quantity) AS max_quantity
-                    FROM priced
-                    GROUP BY date
-                )
-                SELECT
-                    ds.stat_date,
-                    da.min_price,
-                    da.avg_price,
-                    da.p25_price,
-                    da.p75_price,
-                    da.max_price,
-                    da.min_quantity,
-                    da.avg_quantity,
-                    da.max_quantity
-                FROM date_spine ds
-                LEFT JOIN daily_agg da ON da.stat_date = ds.stat_date
-                ORDER BY ds.stat_date
-                """.trimIndent()
-            } else {
-                """
-                WITH RECURSIVE date_spine AS (
-                    SELECT ? AS stat_date
-                    UNION ALL
-                    SELECT DATE_ADD(stat_date, INTERVAL 1 DAY)
-                    FROM date_spine
-                    WHERE stat_date < ?
-                ),
-                priced AS (
-                    SELECT
-                        v.date,
-                        v.price,
-                        v.quantity,
-                        PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY v.price) OVER (
-                            PARTITION BY v.connected_realm_id, v.item_id, v.date
-                        ) AS p25_price,
-                        PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY v.price) OVER (
-                            PARTITION BY v.connected_realm_id, v.item_id, v.date
-                        ) AS p75_price
-                    FROM v_auction_house_prices v
-                    WHERE v.connected_realm_id = ?
-                      AND v.item_id = ?
-                      AND v.date BETWEEN ? AND ?
-                ),
-                daily_agg AS (
-                    SELECT
-                        date AS stat_date,
-                        MIN(price) AS min_price,
-                        AVG(price) AS avg_price,
-                        MIN(p25_price) AS p25_price,
-                        MIN(p75_price) AS p75_price,
-                        MAX(price) AS max_price,
-                        MIN(quantity) AS min_quantity,
-                        AVG(quantity) AS avg_quantity,
-                        MAX(quantity) AS max_quantity
-                    FROM priced
-                    GROUP BY date
-                )
-                SELECT
-                    ds.stat_date,
-                    da.min_price,
-                    da.avg_price,
-                    da.p25_price,
-                    da.p75_price,
-                    da.max_price,
-                    da.min_quantity,
-                    da.avg_quantity,
-                    da.max_quantity
-                FROM date_spine ds
-                LEFT JOIN daily_agg da ON da.stat_date = ds.stat_date
-                ORDER BY ds.stat_date
-                """.trimIndent()
-            }
-        val params: Array<Any?> =
-            if (variant) {
-                arrayOf(
-                    Date.valueOf(fromDate),
-                    Date.valueOf(toDate),
-                    connectedRealmId,
-                    itemId,
-                    Date.valueOf(fromDate),
-                    Date.valueOf(toDate),
-                    bonusKey,
-                    modifierKey,
-                    petSpeciesId,
-                )
-            } else {
-                arrayOf(
-                    Date.valueOf(fromDate),
-                    Date.valueOf(toDate),
-                    connectedRealmId,
-                    itemId,
-                    Date.valueOf(fromDate),
-                    Date.valueOf(toDate),
-                )
-            }
-        return sql to params
-    }
+    ): Pair<String, Array<Any?>> =
+        AuctionMarketItemDetailSql.buildDailySqlAndArgs(
+            connectedRealmId,
+            itemId,
+            fromDate,
+            toDate,
+            variant,
+            bonusKey,
+            modifierKey,
+            petSpeciesId,
+        )
 
     internal fun buildHourlySqlAndArgs(
         connectedRealmId: Int,
@@ -627,129 +341,17 @@ class AuctionMarketItemDetailRepository(
         bonusKey: String,
         modifierKey: String,
         petSpeciesId: Int,
-    ): Pair<String, Array<Any?>> {
-        val priceCase = hourlyPriceCaseExpression("ash", "hs.hour_of_day")
-        val quantityCase = hourlyQuantityCaseExpression("ash", "hs.hour_of_day")
-        val sql =
-            if (variant) {
-                """
-                WITH RECURSIVE date_spine AS (
-                    SELECT ? AS stat_date
-                    UNION ALL
-                    SELECT DATE_ADD(stat_date, INTERVAL 1 DAY)
-                    FROM date_spine
-                    WHERE stat_date < ?
-                ),
-                hour_spine AS (
-                    SELECT 0 AS hour_of_day
-                    UNION ALL
-                    SELECT hour_of_day + 1 FROM hour_spine WHERE hour_of_day < 23
-                ),
-                hourly_agg AS (
-                    SELECT
-                        ds.stat_date,
-                        hs.hour_of_day,
-                        MIN($priceCase) AS min_price,
-                        AVG($priceCase) AS avg_price,
-                        MAX($priceCase) AS max_price,
-                        SUM($quantityCase) AS total_quantity
-                    FROM date_spine ds
-                    CROSS JOIN hour_spine hs
-                    LEFT JOIN auction_stats_hourly ash
-                      ON ash.connected_realm_id = ?
-                     AND ash.item_id = ?
-                     AND ash.date = ds.stat_date
-                     AND ash.bonus_key <=> ?
-                     AND ash.modifier_key <=> ?
-                     AND ash.pet_species_id = ?
-                    GROUP BY ds.stat_date, hs.hour_of_day
-                )
-                SELECT
-                    ds.stat_date,
-                    hs.hour_of_day,
-                    DATE_ADD(TIMESTAMP(ds.stat_date), INTERVAL hs.hour_of_day HOUR) AS timestamp,
-                    ha.min_price,
-                    ha.avg_price,
-                    NULL AS p25_price,
-                    NULL AS p75_price,
-                    ha.max_price,
-                    ha.total_quantity
-                FROM date_spine ds
-                CROSS JOIN hour_spine hs
-                LEFT JOIN hourly_agg ha
-                  ON ha.stat_date = ds.stat_date
-                 AND ha.hour_of_day = hs.hour_of_day
-                ORDER BY ds.stat_date, hs.hour_of_day
-                """.trimIndent()
-            } else {
-                """
-                WITH RECURSIVE date_spine AS (
-                    SELECT ? AS stat_date
-                    UNION ALL
-                    SELECT DATE_ADD(stat_date, INTERVAL 1 DAY)
-                    FROM date_spine
-                    WHERE stat_date < ?
-                ),
-                hour_spine AS (
-                    SELECT 0 AS hour_of_day
-                    UNION ALL
-                    SELECT hour_of_day + 1 FROM hour_spine WHERE hour_of_day < 23
-                ),
-                hourly_agg AS (
-                    SELECT
-                        ds.stat_date,
-                        hs.hour_of_day,
-                        MIN($priceCase) AS min_price,
-                        AVG($priceCase) AS avg_price,
-                        MAX($priceCase) AS max_price,
-                        SUM($quantityCase) AS total_quantity
-                    FROM date_spine ds
-                    CROSS JOIN hour_spine hs
-                    LEFT JOIN auction_stats_hourly ash
-                      ON ash.connected_realm_id = ?
-                     AND ash.item_id = ?
-                     AND ash.date = ds.stat_date
-                    GROUP BY ds.stat_date, hs.hour_of_day
-                )
-                SELECT
-                    ds.stat_date,
-                    hs.hour_of_day,
-                    DATE_ADD(TIMESTAMP(ds.stat_date), INTERVAL hs.hour_of_day HOUR) AS timestamp,
-                    ha.min_price,
-                    ha.avg_price,
-                    NULL AS p25_price,
-                    NULL AS p75_price,
-                    ha.max_price,
-                    ha.total_quantity
-                FROM date_spine ds
-                CROSS JOIN hour_spine hs
-                LEFT JOIN hourly_agg ha
-                  ON ha.stat_date = ds.stat_date
-                 AND ha.hour_of_day = hs.hour_of_day
-                ORDER BY ds.stat_date, hs.hour_of_day
-                """.trimIndent()
-            }
-        val params: Array<Any?> =
-            if (variant) {
-                arrayOf(
-                    Date.valueOf(fromDate),
-                    Date.valueOf(toDate),
-                    connectedRealmId,
-                    itemId,
-                    bonusKey,
-                    modifierKey,
-                    petSpeciesId,
-                )
-            } else {
-                arrayOf(
-                    Date.valueOf(fromDate),
-                    Date.valueOf(toDate),
-                    connectedRealmId,
-                    itemId,
-                )
-            }
-        return sql to params
-    }
+    ): Pair<String, Array<Any?>> =
+        AuctionMarketItemDetailSql.buildHourlySqlAndArgs(
+            connectedRealmId,
+            itemId,
+            fromDate,
+            toDate,
+            variant,
+            bonusKey,
+            modifierKey,
+            petSpeciesId,
+        )
 
     internal fun buildPieSqlAndArgs(
         connectedRealmId: Int,
@@ -759,54 +361,16 @@ class AuctionMarketItemDetailRepository(
         bonusKey: String,
         modifierKey: String,
         petSpeciesId: Int,
-    ): Pair<String, Array<Any?>> {
-        val whereVariant =
-            if (variant) {
-                "AND v.bonus_key <=> ? AND v.modifier_key <=> ? AND v.pet_species_id = ?"
-            } else {
-                ""
-            }
-        val sql =
-            """
-            WITH hours AS (
-                SELECT v.hour_of_day, SUM(v.quantity) AS qty
-                FROM v_auction_house_prices v
-                WHERE v.connected_realm_id = ?
-                  AND v.item_id = ?
-                  AND v.date = ?
-                  $whereVariant
-                GROUP BY v.hour_of_day
-            ),
-            total AS (
-                SELECT COALESCE(SUM(qty), 0) AS t FROM hours
-            )
-            SELECT
-                h.hour_of_day,
-                h.qty AS quantity,
-                CASE WHEN total.t > 0 THEN h.qty / total.t ELSE 0 END AS fraction
-            FROM hours h
-            CROSS JOIN total
-            ORDER BY h.hour_of_day
-            """.trimIndent()
-        val params: Array<Any?> =
-            if (variant) {
-                arrayOf(
-                    connectedRealmId,
-                    itemId,
-                    Date.valueOf(statDate),
-                    bonusKey,
-                    modifierKey,
-                    petSpeciesId,
-                )
-            } else {
-                arrayOf(
-                    connectedRealmId,
-                    itemId,
-                    Date.valueOf(statDate),
-                )
-            }
-        return sql to params
-    }
+    ): Pair<String, Array<Any?>> =
+        AuctionMarketItemDetailSql.buildPieSqlAndArgs(
+            connectedRealmId,
+            itemId,
+            statDate,
+            variant,
+            bonusKey,
+            modifierKey,
+            petSpeciesId,
+        )
 
     internal fun buildCraftingSqlAndArgs(
         connectedRealmId: Int,
@@ -822,176 +386,22 @@ class AuctionMarketItemDetailRepository(
         petSpeciesId: Int,
         preferredRecipeId: Int?,
         localeColumnSuffix: String,
-    ): Pair<String, Array<Any?>> {
-        val hourSuffix = hourColumnSuffix(hourOfDay)
-        val commodityHourSuffix = hourColumnSuffix(commodityHourOfDay)
-        val outputWhere =
-            if (variant) {
-                "AND ash.bonus_key <=> ? AND ash.modifier_key <=> ? AND ash.pet_species_id = ?"
-            } else {
-                ""
-            }
-        val sql =
-            """
-            WITH
-            reagent_sel_base AS (
-                SELECT ash.item_id, ash.price$hourSuffix AS price, ash.bonus_key, ash.modifier_key, ash.pet_species_id
-                FROM auction_stats_hourly ash
-                WHERE ash.connected_realm_id = ?
-                  AND ash.date = ?
-                  AND ash.price$hourSuffix IS NOT NULL
-                  AND ash.item_id IN (SELECT DISTINCT rr.item_id FROM recipe_reagent rr)
-            ),
-            reagent_sel_ranked AS (
-                SELECT item_id, price,
-                       ROW_NUMBER() OVER (PARTITION BY item_id ORDER BY price ASC, bonus_key, modifier_key, pet_species_id) AS rn
-                FROM reagent_sel_base
-            ),
-            reagent_sel AS (
-                SELECT item_id, price FROM reagent_sel_ranked WHERE rn = 1
-            ),
-            reagent_com_base AS (
-                SELECT ash.item_id, ash.price$commodityHourSuffix AS price, ash.bonus_key, ash.modifier_key, ash.pet_species_id
-                FROM auction_stats_hourly ash
-                WHERE ash.connected_realm_id = ?
-                  AND ash.date = ?
-                  AND ash.price$commodityHourSuffix IS NOT NULL
-                  AND ash.item_id IN (SELECT DISTINCT rr.item_id FROM recipe_reagent rr)
-            ),
-            reagent_com_ranked AS (
-                SELECT item_id, price,
-                       ROW_NUMBER() OVER (PARTITION BY item_id ORDER BY price ASC, bonus_key, modifier_key, pet_species_id) AS rn
-                FROM reagent_com_base
-            ),
-            reagent_com AS (
-                SELECT item_id, price FROM reagent_com_ranked WHERE rn = 1
-            ),
-            reagent_price AS (
-                SELECT items.item_id, COALESCE(rs.price, rc.price) AS price
-                FROM (SELECT DISTINCT item_id FROM recipe_reagent) items
-                LEFT JOIN reagent_sel rs ON rs.item_id = items.item_id
-                LEFT JOIN reagent_com rc ON rc.item_id = items.item_id
-            ),
-            recipe_reagent_agg AS (
-                SELECT
-                    r.id AS recipe_id,
-                    SUM(CASE WHEN rr.internal_id IS NULL THEN 0 WHEN rp.price IS NULL THEN 1 ELSE 0 END) AS missing_reagents,
-                    SUM(CASE WHEN rr.internal_id IS NULL THEN 0 ELSE COALESCE(rp.price, 0) * rr.quantity END) AS reagent_cost_partial
-                FROM recipe r
-                LEFT JOIN recipe_reagent rr ON rr.recipe_id = r.id
-                LEFT JOIN reagent_price rp ON rp.item_id = rr.item_id
-                WHERE r.crafted_item_id = ?
-                GROUP BY r.id
-            ),
-            recipe_reagent_cost AS (
-                SELECT
-                    recipe_id,
-                    CASE WHEN missing_reagents > 0 THEN NULL ELSE reagent_cost_partial END AS reagent_cost,
-                    missing_reagents = 0 AS reagents_fully_priced
-                FROM recipe_reagent_agg
-            ),
-            output_sel_base AS (
-                SELECT ash.item_id, ash.price$hourSuffix AS output_unit_price, ash.bonus_key, ash.modifier_key, ash.pet_species_id
-                FROM auction_stats_hourly ash
-                WHERE ash.connected_realm_id = ?
-                  AND ash.date = ?
-                  AND ash.item_id = ?
-                  AND ash.price$hourSuffix IS NOT NULL
-                  $outputWhere
-            ),
-            output_sel_ranked AS (
-                SELECT output_unit_price,
-                       ROW_NUMBER() OVER (ORDER BY output_unit_price ASC, bonus_key, modifier_key, pet_species_id) AS rn
-                FROM output_sel_base
-            ),
-            output_sel AS (
-                SELECT output_unit_price FROM output_sel_ranked WHERE rn = 1
-            ),
-            output_com_base AS (
-                SELECT ash.item_id, ash.price$commodityHourSuffix AS output_unit_price, ash.bonus_key, ash.modifier_key, ash.pet_species_id
-                FROM auction_stats_hourly ash
-                WHERE ash.connected_realm_id = ?
-                  AND ash.date = ?
-                  AND ash.item_id = ?
-                  AND ash.price$commodityHourSuffix IS NOT NULL
-                  $outputWhere
-            ),
-            output_com_ranked AS (
-                SELECT output_unit_price,
-                       ROW_NUMBER() OVER (ORDER BY output_unit_price ASC, bonus_key, modifier_key, pet_species_id) AS rn
-                FROM output_com_base
-            ),
-            output_com AS (
-                SELECT output_unit_price FROM output_com_ranked WHERE rn = 1
-            ),
-            output_price AS (
-                SELECT COALESCE(os.output_unit_price, oc.output_unit_price) AS output_unit_price
-                FROM (SELECT 1 AS id) seed
-                LEFT JOIN output_sel os ON TRUE
-                LEFT JOIN output_com oc ON TRUE
-            )
-            SELECT
-                r.id AS recipe_id,
-                COALESCE(l.$localeColumnSuffix, l.en_gb, l.en_us, CAST(r.id AS CHAR)) AS recipe_name,
-                r.media_url AS recipe_media_url,
-                COALESCE(NULLIF(r.crafted_quantity, 0), 1) AS crafted_quantity,
-                rrc.reagent_cost,
-                COALESCE(rrc.reagents_fully_priced, TRUE) AS reagents_fully_priced,
-                op.output_unit_price,
-                CASE
-                    WHEN rrc.reagents_fully_priced AND rrc.reagent_cost IS NOT NULL AND op.output_unit_price IS NOT NULL
-                    THEN op.output_unit_price * COALESCE(NULLIF(r.crafted_quantity, 0), 1) - rrc.reagent_cost
-                    ELSE NULL
-                END AS profit,
-                CASE
-                    WHEN rrc.reagents_fully_priced AND rrc.reagent_cost IS NOT NULL AND rrc.reagent_cost > 0 AND op.output_unit_price IS NOT NULL
-                    THEN 100.0 * (op.output_unit_price * COALESCE(NULLIF(r.crafted_quantity, 0), 1) - rrc.reagent_cost) / rrc.reagent_cost
-                    ELSE NULL
-                END AS roi_percent
-            FROM recipe r
-            LEFT JOIN recipe_reagent_cost rrc ON rrc.recipe_id = r.id
-            LEFT JOIN locale l ON l.id = r.name_id
-            LEFT JOIN output_price op ON TRUE
-            WHERE r.crafted_item_id = ?
-            ORDER BY (r.id = ?) DESC, (profit IS NULL), profit DESC, r.id
-            """.trimIndent()
-        val outputParams: Array<Any?> =
-            if (variant) {
-                arrayOf(
-                    connectedRealmId,
-                    Date.valueOf(statDate),
-                    itemId,
-                    bonusKey,
-                    modifierKey,
-                    petSpeciesId,
-                    commodityConnectedRealmId,
-                    Date.valueOf(commodityStatDate),
-                    itemId,
-                    bonusKey,
-                    modifierKey,
-                    petSpeciesId,
-                )
-            } else {
-                arrayOf(
-                    connectedRealmId,
-                    Date.valueOf(statDate),
-                    itemId,
-                    commodityConnectedRealmId,
-                    Date.valueOf(commodityStatDate),
-                    itemId,
-                )
-            }
-        return sql to arrayOf(
+    ): Pair<String, Array<Any?>> =
+        AuctionMarketItemDetailSql.buildCraftingSqlAndArgs(
             connectedRealmId,
-            Date.valueOf(statDate),
             commodityConnectedRealmId,
-            Date.valueOf(commodityStatDate),
             itemId,
-            *outputParams,
-            itemId,
-            preferredRecipeId ?: -1,
+            statDate,
+            commodityStatDate,
+            hourOfDay,
+            commodityHourOfDay,
+            variant,
+            bonusKey,
+            modifierKey,
+            petSpeciesId,
+            preferredRecipeId,
+            localeColumnSuffix,
         )
-    }
 
     /**
      * Returns true when the given recipe exists and produces the given crafted item. Used to enforce
@@ -1144,7 +554,7 @@ class AuctionMarketItemDetailRepository(
         val outputParams: Array<Any?> = if (variant) arrayOf(bonusKey, modifierKey, petSpeciesId) else emptyArray()
         return jdbcTemplate.query(
             sql,
-            craftingAnalyticsDailyRowMapper,
+            AuctionMarketItemDetailRowMappers.craftingAnalyticsDailyRowMapper,
             Date.valueOf(fromDate),
             Date.valueOf(toDate),
             connectedRealmId,
@@ -1353,7 +763,7 @@ class AuctionMarketItemDetailRepository(
         val variantParams: Array<Any?> = if (variant) arrayOf(bonusKey, modifierKey, petSpeciesId) else emptyArray()
         return jdbcTemplate.query(
             sql,
-            craftingAnalyticsHeatmapRowMapper,
+            AuctionMarketItemDetailRowMappers.craftingAnalyticsHeatmapRowMapper,
             Date.valueOf(fromDate),
             Date.valueOf(toDate),
             connectedRealmId,
@@ -1380,29 +790,6 @@ class AuctionMarketItemDetailRepository(
             itemId,
         )
     }
-
-    private val craftingAnalyticsDailyRowMapper =
-        RowMapper { rs: ResultSet, _: Int ->
-            AuctionMarketItemCraftingAnalyticsDailyRow(
-                statDate = rs.getObject("stat_date", LocalDate::class.java),
-                profit = rs.getNullableLong("profit"),
-                roiPercent = rs.getNullableDouble("roi_percent"),
-                reagentCost = rs.getNullableLong("reagent_cost"),
-                outputUnitPrice = rs.getNullableLong("output_unit_price"),
-            )
-        }
-
-    private val craftingAnalyticsHeatmapRowMapper =
-        RowMapper { rs: ResultSet, _: Int ->
-            AuctionMarketItemCraftingHeatmapRow(
-                dayOfWeek = rs.getInt("day_of_week"),
-                hourOfDay = rs.getInt("hour_of_day"),
-                profit = rs.getNullableDouble("profit"),
-                outputUnitPrice = rs.getNullableDouble("output_unit_price"),
-                roiPercent = rs.getNullableDouble("roi_percent"),
-                sampleCount = rs.getInt("sample_count"),
-            )
-        }
 
     private fun hourColumnSuffix(hourOfDay: Int): String = hourOfDay.coerceIn(0, 23).toString().padStart(2, '0')
